Created Rscript for project.
I have started the Rscript and have made good progress. It does converge and seem to work, but I think I may need to add another column for the intercepts for X matrix so I will continue working and let you all know when I feel it is complete for you all to review.